requirements.txt
requests
beautifulsoup4
markdownify
python-slugify
google-genai
python-dotenv